# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local Business Support Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterUserBodyRole =
  (typeof RegisterUserBodyRole)[keyof typeof RegisterUserBodyRole];

export const RegisterUserBodyRole = {
  customer: "customer",
  shop_owner: "shop_owner",
} as const;

export interface RegisterUserBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
  role: RegisterUserBodyRole;
  /** Required if role is shop_owner */
  businessName?: string;
  locality?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  shop_owner: "shop_owner",
  admin: "admin",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: UserRole;
  status: UserStatus;
  businessName?: string | null;
  locality?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ShopCategory = (typeof ShopCategory)[keyof typeof ShopCategory];

export const ShopCategory = {
  grocery: "grocery",
  fashion: "fashion",
  healthcare: "healthcare",
  services: "services",
  restaurants: "restaurants",
} as const;

export type ShopStatus = (typeof ShopStatus)[keyof typeof ShopStatus];

export const ShopStatus = {
  pending: "pending",
  verified: "verified",
  rejected: "rejected",
} as const;

export interface Shop {
  id: number;
  name: string;
  category: ShopCategory;
  /** For healthcare shops - hospital, clinic, pharmacy, diagnostic, optical */
  subcategory?: string | null;
  locality: string;
  phone?: string | null;
  imageUrl?: string | null;
  openingTime?: string | null;
  closingTime?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  status: ShopStatus;
  ownerId?: number | null;
  rating?: number | null;
  reviewCount?: number | null;
  createdAt: string;
}

export type CreateShopBodyCategory =
  (typeof CreateShopBodyCategory)[keyof typeof CreateShopBodyCategory];

export const CreateShopBodyCategory = {
  grocery: "grocery",
  fashion: "fashion",
  healthcare: "healthcare",
  services: "services",
  restaurants: "restaurants",
} as const;

export interface CreateShopBody {
  name: string;
  category: CreateShopBodyCategory;
  subcategory?: string | null;
  locality: string;
  phone?: string | null;
  imageUrl?: string | null;
  openingTime?: string | null;
  closingTime?: string | null;
}

export interface UpdateShopBody {
  name?: string;
  locality?: string;
  phone?: string | null;
  subcategory?: string | null;
  openingTime?: string | null;
  closingTime?: string | null;
}

export interface Medicine {
  id: number;
  shopId: number;
  name: string;
  available: boolean;
  price?: number | null;
}

export type ShopsSummaryByCategory = { [key: string]: number };

export interface ShopsSummary {
  total: number;
  byCategory: ShopsSummaryByCategory;
  localities: string[];
}

export type VerifyUserBodyAction =
  (typeof VerifyUserBodyAction)[keyof typeof VerifyUserBodyAction];

export const VerifyUserBodyAction = {
  approve: "approve",
  reject: "reject",
} as const;

export interface VerifyUserBody {
  action: VerifyUserBodyAction;
  reason?: string | null;
}

export type VerifyShopBodyAction =
  (typeof VerifyShopBodyAction)[keyof typeof VerifyShopBodyAction];

export const VerifyShopBodyAction = {
  approve: "approve",
  reject: "reject",
} as const;

export interface VerifyShopBody {
  action: VerifyShopBodyAction;
  reason?: string | null;
}

export interface AdminStats {
  pendingUsers: number;
  pendingShops: number;
  totalUsers: number;
  totalShops: number;
  totalCustomers: number;
  totalShopOwners: number;
  verifiedShops: number;
}

export interface Product {
  id: number;
  shopId: number;
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  available: boolean;
  category?: string | null;
}

export interface CreateProductBody {
  name: string;
  description?: string | null;
  price: number;
  imageUrl?: string | null;
  available?: boolean;
  category?: string | null;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  quantity: number;
  priceAtTime: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  shopId: number;
  totalAmount: number;
  status: OrderStatus;
  deliveryAddress: string;
  createdAt: string;
  items?: OrderItem[];
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
};

export interface CreateOrderBody {
  shopId: number;
  deliveryAddress: string;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type ListShopsParams = {
  category?: string;
  search?: string;
  locality?: string;
};

export type ListHealthcareShopsParams = {
  subcategory?: ListHealthcareShopsSubcategory;
  search?: string;
};

export type ListHealthcareShopsSubcategory =
  (typeof ListHealthcareShopsSubcategory)[keyof typeof ListHealthcareShopsSubcategory];

export const ListHealthcareShopsSubcategory = {
  hospital: "hospital",
  clinic: "clinic",
  pharmacy: "pharmacy",
  diagnostic: "diagnostic",
  optical: "optical",
} as const;